EQUIVARIANT_NETWORKS_PLANETARY_SYSTEMS_ARCHITECTURES/midterm_submission/requirements.txt
torch
torchvision
pytorch-lightning
e3nn
wandb
pandas
numpy
Pillow
e2cnn